Revert "Added BarAnimation Component"
