+++ frontend/package.json
@@ -15,7 +15,6 @@
     "framer-motion": "^12.4.7",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-router-dom": "^7.2.0",
     "tailwindcss": "^4.0.9"
   },
   "devDependencies": {
