+++ frontend/src/App.jsx
@@ -1,13 +1,9 @@
-import React from "react";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Welcomepage from "./Components/WelcomePage";
-
+import React from 'react'
+import WelcomePage from './Components/WelcomePage'
 export default function App() {
   return (
-    <Router>
-      <Routes>
-        <Route path="/" element={<Welcomepage />} />
-      </Routes>
-    </Router>
-  );
+    <div>
+      <WelcomePage />
+    </div>
+  )
 }
